+++ components/org.wso2.carbon.analytics.msf4j.interceptor.common/src/main/java/org/wso2/carbon/analytics/msf4j/interceptor/common/config/SPInterceptorConfig.java
@@ -15,31 +15,27 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.wso2.carbon.msf4j.interceptor.common.internal;
 
-import org.wso2.carbon.stream.processor.idp.client.core.api.IdPClient;
+package org.wso2.carbon.analytics.msf4j.interceptor.common.config;
 
-public class DataHolder {
-    private static DataHolder instance = new DataHolder();
-    private IdPClient idPClient;
+import org.osgi.framework.BundleContext;
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+import org.wso2.msf4j.interceptor.OSGiInterceptorConfig;
 
-    private DataHolder() {
-    }
 
-    /**
-     * Provide instance of DataHolder class.
-     *
-     * @return Instance of DataHolder
-     */
-    public static DataHolder getInstance() {
-        return instance;
-    }
+/**
+ * Interceptor configuration for SiddhiAppAPI.
+ */
 
-    public IdPClient getIdPClient() {
-        return idPClient;
-    }
+@Component(
+        name = "stream-processor-interceptor-config",
+        service = OSGiInterceptorConfig.class,
+        immediate = true
+)
+public class SPInterceptorConfig extends OSGiInterceptorConfig {
 
-    public void setIdPClient(IdPClient idPClient) {
-        this.idPClient = idPClient;
+    @Activate
+    protected void activate(BundleContext bundleContext) {
     }
 }
