+++ components/org.wso2.carbon.stream.processor.auth.rest.api/src/main/java/org/wso2/carbon/stream/processor/auth/rest/api/internal/ServiceComponent.java
@@ -52,7 +52,6 @@ public class ServiceComponent {
         DataHolder.getInstance().setIdPClient(client);
     }
 
-
     protected void unregisterIdP(IdPClient client) {
     }
 }
