+++ features/org.wso2.carbon.analytics.msf4j.interceptor.common.feature/pom.xml
@@ -26,7 +26,7 @@
         <relativePath>../../pom.xml</relativePath>
     </parent>
 
-    <artifactId>org.wso2.carbon.msf4j.interceptor.common.feature</artifactId>
+    <artifactId>org.wso2.carbon.analytics.msf4j.interceptor.common.feature</artifactId>
     <packaging>carbon-feature</packaging>
     <name>WSO2 Data Analytics Server - Interceptor Common Feature</name>
     <url>http://wso2.org</url>
@@ -35,7 +35,7 @@
     <dependencies>
         <dependency>
             <groupId>org.wso2.carbon.analytics</groupId>
-            <artifactId>org.wso2.carbon.msf4j.interceptor.common</artifactId>
+            <artifactId>org.wso2.carbon.analytics.msf4j.interceptor.common</artifactId>
         </dependency>
     </dependencies>
 
@@ -66,7 +66,7 @@
                             </adviceFileContents>
                             <bundles>
                                 <bundle>
-                                    <symbolicName>org.wso2.carbon.msf4j.interceptor.common</symbolicName>
+                                    <symbolicName>org.wso2.carbon.analytics.msf4j.interceptor.common</symbolicName>
                                     <version>${project.version}</version>
                                 </bundle>
                             </bundles>
